Add files via upload
XBT is a program that makes keeping all of your user data safely backed up on
a dedicated External USB drive easy. XBT works with Ubuntu 16.04 onward and
the Linux Mint 18.x series.

XBT will update files that have changed, remove files that have been
deleted and add any new files that have been created since the last backup.
The initial backup can take a lot of time if you have a lot of data stored
in your system's /home directory. XBT creates a directory with the host name
of the machine it's backing up from inside XBT-Backups. This allows users to
share one XBT_Drive with multiple machines.

The directory structure and all files are stored openly to allow users easy
access if they only need to retrieve a few files or directories using a file
manager. Data from /home can be restored with the built in "xbt restore"
command if the user wants to roll back to the exact state of /home when the
last backup was taken.